requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=1.5.3
numpy>=1.24.3

# Document parsing
python-docx>=0.8.11
PyPDF2>=3.0.1
openpyxl>=3.1.2

# NLP and text processing
spacy>=3.5.3
nltk>=3.8.1
sentence-transformers>=2.2.2

# LLM API clients
openai>=1.1.1
anthropic>=0.5.0
google-generativeai>=0.2.0

# Development tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0
mkdocs>=1.4.3 